+++ docs/release-notes.md
@@ -1,10 +1,26 @@
+## 0.37.1
+
+February 9, 2024
+
+#### Fixed
+
+* Warn instead of raise for missing env file on `Config` [#2485](https://github.com/encode/starlette/pull/2485).
+
+## 0.37.0
+
+February 5, 2024
+
+#### Added
+
+* Support the WebSocket Denial Response ASGI extension [#2041](https://github.com/encode/starlette/pull/2041).
+
 ## 0.36.3
 
 February 4, 2024
 
 #### Fixed
 
-* Create `anyio.Event` on async context [#2459](https://github.com/encode/starlette/2459).
+* Create `anyio.Event` on async context [#2459](https://github.com/encode/starlette/pull/2459).
 
 ## 0.36.2